requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
pymongo==4.6.0
python-socketio==5.10.0 